refactor to chart widgets
