+++ src/widget/chart/mod.rs
@@ -1,0 +1,5 @@
+pub use self::prices_line::PricesLineChart;
+pub use self::volume_bar::VolumeBarChart;
+
+mod prices_line;
+mod volume_bar;
