+++ src/widget/stock_summary.rs
@@ -1,18 +1,14 @@
-use itertools::Itertools;
 use tui::buffer::Buffer;
 use tui::layout::{Alignment, Constraint, Direction, Layout, Rect};
 use tui::style::{Modifier, Style};
-use tui::symbols::{bar, Marker};
 use tui::text::{Span, Spans};
-use tui::widgets::{
-    Axis, BarChart, Block, Borders, Chart, Dataset, GraphType, Paragraph, StatefulWidget, Widget,
-};
+use tui::widgets::{Block, Borders, Paragraph, StatefulWidget, Widget};
 
+use super::chart::{PricesLineChart, VolumeBarChart};
 use super::stock::StockState;
 use super::{CachableWidget, CacheState};
-use crate::common::*;
 use crate::draw::{add_padding, PaddingDirection};
-use crate::{ENABLE_PRE_POST, HIDE_PREV_CLOSE, SHOW_VOLUMES, THEME};
+use crate::{ENABLE_PRE_POST, SHOW_VOLUMES, THEME};
 
 pub struct StockSummaryWidget {}
 
@@ -163,246 +159,37 @@ impl CachableWidget<StockState> for StockSummaryWidget {
                 .render(layout[0], buf);
         }
 
-        // Draw graph
-        {
-            layout[1] = add_padding(layout[1], 1, PaddingDirection::Left);
+        // graph_chunks[0] = prices
+        // graph_chunks[1] = volume
+        let graph_chunks = if show_volumes {
+            Layout::default()
+                .constraints([Constraint::Min(5), Constraint::Length(1)].as_ref())
+                .split(layout[1])
+        } else {
+            Layout::default()
+                .constraints([Constraint::Min(0)].as_ref())
+                .split(layout[1])
+        };
 
-            let (min, max) = state.min_max(&data);
-            let (start, end) = state.start_end();
+        // Draw prices line chart
+        PricesLineChart {
+            data: &data,
+            enable_pre_post,
+            is_profit: pct_change >= 0.0,
+            is_summary: true,
+            loaded,
+            show_x_labels: false,
+        }
+        .render(graph_chunks[0], buf, state);
 
-            let mut prices: Vec<_> = data.iter().map(cast_historical_as_price).collect();
-
-            prices.pop();
-            prices.push(state.current_price());
-            zeros_as_pre(&mut prices);
-
-            // Need more than one price for GraphType::Line to work
-            let graph_type = if prices.len() <= 2 {
-                GraphType::Scatter
-            } else {
-                GraphType::Line
-            };
-
-            let trading_period = state.current_trading_period(&data);
-
-            let (reg_prices, pre_prices, post_prices) = if loaded {
-                let (start_idx, end_idx) = state.regular_start_end_idx(&data);
-
-                if enable_pre_post && state.time_frame == TimeFrame::Day1 {
-                    (
-                        prices
-                            .iter()
-                            .enumerate()
-                            .filter(|(idx, _)| {
-                                if let Some(start) = start_idx {
-                                    *idx >= start
-                                } else {
-                                    false
-                                }
-                            })
-                            .filter(|(idx, _)| {
-                                if let Some(end) = end_idx {
-                                    *idx <= end
-                                } else {
-                                    true
-                                }
-                            })
-                            .map(cast_as_dataset)
-                            .collect::<Vec<(f64, f64)>>(),
-                        {
-                            let pre_end_idx = if let Some(start_idx) = start_idx {
-                                start_idx
-                            } else {
-                                prices.len()
-                            };
-
-                            if pre_end_idx > 0 {
-                                Some(
-                                    prices
-                                        .iter()
-                                        .enumerate()
-                                        .filter(|(idx, _)| *idx <= pre_end_idx)
-                                        .map(cast_as_dataset)
-                                        .collect::<Vec<(f64, f64)>>(),
-                                )
-                            } else {
-                                None
-                            }
-                        },
-                        {
-                            if let Some(post_start_idx) = end_idx {
-                                Some(
-                                    prices
-                                        .iter()
-                                        .enumerate()
-                                        .filter(|(idx, _)| *idx >= post_start_idx)
-                                        .map(cast_as_dataset)
-                                        .collect::<Vec<(f64, f64)>>(),
-                                )
-                            } else {
-                                None
-                            }
-                        },
-                    )
-                } else {
-                    (
-                        prices
-                            .iter()
-                            .enumerate()
-                            .map(cast_as_dataset)
-                            .collect::<Vec<(f64, f64)>>(),
-                        None,
-                        None,
-                    )
-                }
-            } else {
-                (vec![], None, None)
-            };
-
-            let prev_close_line = if state.time_frame == TimeFrame::Day1
-                && loaded
-                && !*HIDE_PREV_CLOSE
-                && state.prev_close_price.is_some()
-            {
-                let num_points = (end - start) / 60 + 1;
-
-                Some(
-                    (0..num_points)
-                        .map(|i| ((i + 1) as f64, state.prev_close_price.unwrap()))
-                        .collect::<Vec<_>>(),
-                )
-            } else {
-                None
-            };
-
-            let mut datasets = vec![Dataset::default()
-                .marker(Marker::Braille)
-                .style(Style::default().fg(
-                    if trading_period != TradingPeriod::Regular && enable_pre_post {
-                        THEME.gray
-                    } else if pct_change >= 0.0 {
-                        THEME.profit
-                    } else {
-                        THEME.loss
-                    },
-                ))
-                .graph_type(graph_type)
-                .data(&reg_prices)];
-
-            if let Some(data) = post_prices.as_ref() {
-                datasets.push(
-                    Dataset::default()
-                        .marker(Marker::Braille)
-                        .style(
-                            Style::default().fg(if trading_period != TradingPeriod::Post {
-                                THEME.gray
-                            } else if pct_change >= 0.0 {
-                                THEME.profit
-                            } else {
-                                THEME.loss
-                            }),
-                        )
-                        .graph_type(GraphType::Line)
-                        .data(&data),
-                );
+        // Draw volumes bar chart
+        if show_volumes {
+            VolumeBarChart {
+                data: &data,
+                loaded,
+                show_x_labels: false,
             }
-
-            if let Some(data) = pre_prices.as_ref() {
-                datasets.insert(
-                    0,
-                    Dataset::default()
-                        .marker(Marker::Braille)
-                        .style(
-                            Style::default().fg(if trading_period != TradingPeriod::Pre {
-                                THEME.gray
-                            } else if pct_change >= 0.0 {
-                                THEME.profit
-                            } else {
-                                THEME.loss
-                            }),
-                        )
-                        .graph_type(GraphType::Line)
-                        .data(&data),
-                );
-            }
-
-            if let Some(data) = prev_close_line.as_ref() {
-                datasets.insert(
-                    0,
-                    Dataset::default()
-                        .marker(Marker::Braille)
-                        .style(Style::default().fg(THEME.gray))
-                        .graph_type(GraphType::Line)
-                        .data(&data),
-                );
-            }
-
-            // graph_chunks[0] = prices
-            // graph_chunks[1] = volume
-            let graph_chunks = if show_volumes {
-                Layout::default()
-                    .constraints([Constraint::Length(5), Constraint::Length(1)].as_ref())
-                    .split(layout[1])
-            } else {
-                vec![layout[1]]
-            };
-
-            if show_volumes {
-                let mut volume_chunks = graph_chunks[1];
-                volume_chunks.height += 1;
-
-                let x_offset = if !loaded { 8 } else { 9 };
-                volume_chunks.x += x_offset;
-
-                if volume_chunks.width > x_offset + 1 {
-                    volume_chunks.width -= x_offset + 1;
-
-                    let width = volume_chunks.width;
-                    let num_bars = width as usize;
-
-                    let volumes = state.volumes(&data);
-                    let vol_count = volumes.len();
-
-                    if vol_count > 0 {
-                        let volumes = data
-                            .iter()
-                            .map(|p| [p.volume].repeat(num_bars))
-                            .flatten()
-                            .chunks(vol_count)
-                            .into_iter()
-                            .map(|c| ("", c.sum::<u64>() / vol_count as u64))
-                            .collect::<Vec<_>>();
-
-                        volume_chunks.x -= 1;
-
-                        Block::default()
-                            .borders(Borders::LEFT)
-                            .border_style(Style::default().fg(THEME.border_axis))
-                            .render(volume_chunks, buf);
-
-                        volume_chunks.x += 1;
-
-                        BarChart::default()
-                            .bar_gap(0)
-                            .bar_set(bar::NINE_LEVELS)
-                            .style(Style::default().fg(THEME.gray))
-                            .data(&volumes)
-                            .render(volume_chunks, buf);
-                    }
-                }
-            }
-
-            Chart::new(datasets)
-                .style(Style::default().bg(THEME.background()))
-                .block(Block::default().border_style(Style::default()))
-                .x_axis(Axis::default().bounds(state.x_bounds(start, end, &data)))
-                .y_axis(
-                    Axis::default()
-                        .bounds(state.y_bounds(min, max))
-                        .labels(state.y_labels(min, max))
-                        .style(Style::default().fg(THEME.border_axis)),
-                )
-                .render(graph_chunks[0], buf);
+            .render(graph_chunks[1], buf, state);
         }
     }
 }
